requirements.txt
appdirs==1.4.0
botocore==1.5.7
codeclimate-test-reporter==0.2.0
coverage==4.3.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.5
django-storages==1.5.2
django-taggit==0.22.0
django-toolbelt==0.0.1
djangorestframework==3.5.3
docutils==0.13.1
gunicorn==19.6.0
jmespath==0.9.1
model-mommy==1.3.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
requests==2.12.4
s3transfer==0.1.10
six==1.10.0
static3==0.7.0
whitenoise==2.0.6
